move tree node type from json to crdt
